+++ src/st_stuff.c
@@ -1507,7 +1507,7 @@ static void ST_LoadUnloadGraphics(load_callback_t callback)
 static void ST_LoadCallback(char *lumpname, patch_t **variable)
 {
     if (M_StringCompare(lumpname, "STARMS") || M_StringCompare(lumpname, "STBAR") || M_StringCompare(lumpname, "STFGOD0"))
-        *variable = ((FREEDOOM && !modifiedgame) || hacx ? W_CacheLastLumpName(lumpname) : W_CacheLumpName(lumpname));
+        *variable = ((FREEDOOM && !modifiedgame) || hacx || REKKR ? W_CacheLastLumpName(lumpname) : W_CacheLumpName(lumpname));
     else
         *variable = W_CacheLumpName(lumpname);
 }